Convert switch to if-else 2 files
With the newer ADT plugins, the generated R file no longer uses 'final' on
all of its resources. Because of this, switch-case causes a compile time
error. This commit simply converts the switch-case statements in 2 files
to if-else statements which work just as well, but don't break when using
ADT to generate the R file.

For further information about this issue, see:
http://tools.android.com/tips/non-constant-fields